+++ lib/src/main/java/com/wjduquette/joe/Interpreter.java
@@ -134,6 +134,35 @@ public class Interpreter {
                     evaluate(stmt.incr());
                 }
             }
+            case Stmt.ForEach stmt -> {
+                var list = evaluate(stmt.listExpr());
+                Collection<?> collection = switch (list) {
+                    case Collection<?> c -> c;
+                    case JoeIterable i -> i.getItems();
+                    default -> {
+                        var instance = joe.getJoeObject(list);
+                        if (instance.canIterate()) {
+                            yield instance.getItems();
+                        } else {
+                            throw new RuntimeError(stmt.varName(),
+                                "Expected iterable, got: " +
+                                    joe.typeName(list) + "'" +
+                                    joe.stringify(list) + "'");
+                        }
+                    }
+                };
+
+                for (var item : collection) {
+                    try {
+                        environment.define(stmt.varName().lexeme(), item);
+                        execute(stmt.body());
+                    } catch (Break ex) {
+                        break;
+                    } catch (Continue ex) {
+                        // Nothing else to do
+                    }
+                }
+            }
             case Stmt.Function stmt -> {
                 var function = new JoeFunction(stmt, environment, false);
                 environment.define(stmt.name().lexeme(), function);