+++ lib/src/main/java/com/wjduquette/joe/JoeIterable.java
@@ -1,0 +1,11 @@
+package com.wjduquette.joe;
+
+import java.util.Collection;
+
+/**
+ * Any object that can return a collection of items can be iterated over
+ * using Joe's `foreach` statement.
+ */
+public interface JoeIterable {
+    Collection<Object> getItems();
+}